packages config
