--- app/src/main/java/com/example/newsportal/MainActivity.kt
@@ -1,21 +1,0 @@
-package com.example.newsportal
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import androidx.fragment.app.commit
-import com.example.newsportal.view.TopHeadlineFragment
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        if (savedInstanceState == null) {
-            supportFragmentManager.commit {
-                setReorderingAllowed(true)
-
-                add(R.id.fragment_container_view, TopHeadlineFragment())
-            }
-        }
-    }
-}
