+++ app/src/main/java/com/example/newsportal/app/main/MainActivity.kt
@@ -1,0 +1,22 @@
+package com.example.newsportal.app.main
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.fragment.app.commit
+import com.example.newsportal.R
+import com.example.newsportal.app.topnews.TopNewsFragment
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        if (savedInstanceState == null) {
+            supportFragmentManager.commit {
+                setReorderingAllowed(true)
+
+                add(R.id.fragment_container_view, TopNewsFragment())
+            }
+        }
+    }
+}
