+++ app/src/main/java/com/example/newsportal/data/local/Database.kt
@@ -1,8 +1,8 @@
-package com.example.newsportal.room
+package com.example.newsportal.data.local
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.example.newsportal.model.ArticleLocal
+import com.example.newsportal.data.local.model.ArticleLocal
 
 @Database(entities = [ArticleLocal::class], version = 1)
 abstract class Database : RoomDatabase(){
