+++ app/src/main/java/com/example/newsportal/data/remote/model/Source.kt
@@ -1,3 +1,3 @@
-package com.example.newsportal.model
+package com.example.newsportal.data.remote.model
 
 data class Source(val id: String, val name: String)
