+++ app/src/main/java/com/example/newsportal/domain/model/Article.kt
@@ -1,0 +1,3 @@
+package com.example.newsportal.domain.model
+
+data class Article(val title: String)
