--- app/src/main/java/com/example/newsportal/model/NewsResponce.kt
@@ -1,3 +1,0 @@
-package com.example.newsportal.model
-
-data class NewsResponce(val status: String, val totalResults: Int, val articles: List<Article>)
