+++ app/src/main/res/layout/fragment_category.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.CategoryFragment">
+    tools:context=".app.topnews.CategoryFragment">
 
     <TextView
         android:id="@+id/tv"
